requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
alibabacloud-credentials==0.2.0
alibabacloud-dysmsapi20170525==2.0.7
alibabacloud-endpoint-util==0.0.3
alibabacloud-openapi-util==0.1.5
alibabacloud-tea==0.2.8
alibabacloud-tea-openapi==0.2.8
alibabacloud-tea-util==0.3.5
alipay-sdk-python==3.6.28
amqp==5.0.6
aniso8601==7.0.0
apispec==5.1.1
apispec-webframeworks==0.5.2
astroid==2.8.5
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
bidict==0.21.4
billiard==3.6.4.0
black==21.11b1
blinker==1.4
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickhouse-cityhash==1.0.2.3
clickhouse-driver==0.2.2
clickhouse-sqlalchemy==0.1.7
colorama==0.4.4
cryptography==35.0.0
distlib==0.3.3
dnspython==2.1.0
docopt==0.6.2
docutils==0.17.1
email-validator==1.1.3
fett==0.3.2
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-GraphQL==2.0.1
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
frozenlist==1.2.0
graphene==2.1.9
graphene-mongo==0.2.13
graphene-sqlalchemy==2.3.0
graphene-sqlalchemy-filter==1.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
infi.clickhouse-orm==2.1.1
iso8601==1.0.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
kafka-python==2.0.2
kombu==5.2.2
lazy-object-proxy==1.6.0
lz4==3.1.10
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-mongoengine==0.30.2
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
minio==7.1.2
mongoengine==0.23.1
multidict==5.2.0
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.4
packaging==21.2
pandas==1.3.4
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
promise==2.3
prompt-toolkit==3.0.22
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.10.1
PyJWT==2.3.0
pylint==2.11.1
pymongo==3.12.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.0
python-jsonrpc-server==0.3.4
python-socketio==5.5.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==3.5.3
regex==2021.11.10
requests==2.26.0
rsa==4.7.2
rstcheck==3.3.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
snooty-lextudio==1.11.6
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-bulk-lazy-loader==0.9.9
toml==0.10.2
tomli==1.2.2
tox==3.24.4
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
watchdog==1.0.2
wcwidth==0.2.5
wechatpayv3==1.2.11
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
yarl==1.7.2
zstd==1.5.0.2
